Add files via upload
The above attached is the Business Analyst Capstone Project of E-Fixed Deposit of ABC Bank. It has file of Business Analysis of ABC Bank,user stories and acceptance criteria of Fixed Deposits in JIRA through Agile Scrum Concept,Dashboard of FD's in Tableau and sales analysis through Bar Graph of FD's in excel.